refactored api calls for categories
